Main

+++ frontend/src/pages/EnhancedFeesPage.tsx
@@ -430,7 +430,7 @@ export const EnhancedFeesPage: React.FC = () => {
   const formatMonthLabel = (month: string) => {
     const [year, monthNum] = month.split('-');
     const date = new Date(parseInt(year), parseInt(monthNum) - 1);
-    return date.toLocaleDateString('en-US', { month: 'long', year: 'numeric' });
+    return date.toLocaleDateString('en-US', { month: 'short', year: 'numeric' });
   };
 
   return (
